// src/api/openapi/model/templatePaint.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Unit } from './unit';


export interface TemplatePaint { 
    id: number;
    name: string;
    price: number;
    unit: Unit;
}